Rename jira_scraper to data_scraper
... and remove unused log_parser.
