+++ data_scraper/core/scraper.py
@@ -10,9 +10,9 @@ import pandas as pd
 from tqdm import tqdm
 from openai import OpenAI
 
-from jira_scraper.processors.jira_provider import JiraProvider
-from jira_scraper.processors.vector_store import QdrantVectorStoreManager
-from jira_scraper.processors.text_processor import TextProcessor
+from data_scraper.processors.jira_provider import JiraProvider
+from data_scraper.processors.vector_store import QdrantVectorStoreManager
+from data_scraper.processors.text_processor import TextProcessor
 
 
 LOG = logging.getLogger(__name__)
