+++ tox.ini
@@ -18,10 +18,10 @@ commands =
 description = run Pylint
 commands =
     {[testenv]commands}
-    pylint {posargs:./jira_scraper}
+    pylint {posargs:./data_scraper}
 
 [testenv:ruff]
 description = run ruff
 commands =
     {[testenv]commands}
-    ruff check {posargs:./jira_scraper}
+    ruff check {posargs:./data_scraper}
